requirements.txt
faiss-cpu==1.7.3
transformers==4.28.1
torch==2.0.0
sentence-transformers==2.2.0
llama_index==0.5.6
pdfplumber==0.8.0
python-docx==0.8.11
pandas==1.5.3
rouge-score==0.1.2
requests==2.28.2
openai==0.27.0
spacy==3.8.2
textwrap==0.1.1
